arreglo primer punto

+++ src/main/java/co/edu/ufps/DTOs/PagoRequest.java
@@ -13,5 +13,5 @@ public class PagoRequest {
 	@JsonProperty("tipo_tarjeta")
     private String tipoTarjeta;
     private int cuotas;
-    private BigDecimal valor;
+    private Integer valor;
 }
